requirements.txt
Flask
tensorflow
numpy
opencv-python-headless
Pillow
fpdf
werkzeug